Don't create instances with DefaultSignatures
We still support GHC 7.0, which doesn't have -XDefaultSignatures
enabled. As a result, we currently have some CPP pragma shims to use
DefaultSignatures for some instances on GHC 7.1+, but not to use
DefaultSignatures on GHc 7.0. This results in a lot of redundancy and
won't scale well if we fix #24, so we will simply implement each
instance manually for now.

If we later drop support for GHC 7.0, going back to DefaultSignatures
would be as simple as removing the manual definitions.