Refactor code in Generics.Deriving.TH
This performs a raft of refactorings in `Generics.Deriving.TH`. Of note is
that the misleadingly-named `GenericKind` is now named `GenericTvbs`, as that
is really what it is about. I also plumb `GenericTvbs` to more places instead
of just `GenericClass`, which enables some further code simplifications.

There is no change in behavior, but this will make a subsequent commit easier
to implement.