Don't bother building a TypeSubst for derive- methods
The `TypeSubst` that we plumb down to `repField` ultimately serves to
replace the reified type variables from `reifyDatatype` with the actual
type variables that the user supplies. For `make-` functions, the user-supplied
type variables can be different, so using a `TypeSubst` is critical. For
`derive-` functions, however, the domain and range of the substitution are
exactly the same, as they both use reified type variables. Therefore, there
is really no need to build up a substitution at all.

In fact, #80 uncovered a bug in which this substitution was built up
incorrectly, leading to subtle code generation bugs. The easiest fix by far is
to just pass an empty substitution down to `repField`, which avoids the issue
altogether in the common case of `derive-` functions. I'm not 100% convinced
that there aren't still some lingering issues in the lesser-used `make-` code
paths, but I'm disinclined to investigate those unless someone specifically
complains about it.

Fixes #80.